requirements.txt
django==3.0.5
djangorestframework==3.11.0